fixed padding issue
in some cases, video_data2 contains non-consecutive datetime indices (so far only encountered once) which screws up the getting of global_first and _last timestamps, and thus padding.
Fixed by workaround - assumes it is the second datetime index what is weird, so simply removes the first. Works for this datasaet, but may come back in the future.